# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravelooo API",
        "contact": {
            "email": "mikahoy045@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pages": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "List all published pages",
                "operationId": "cd59deb18bfa6add22005d63640ccd06",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Page"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pages"
                ],
                "summary": "Create new page",
                "operationId": "22ffe9506aba1ed019b4d54f3d7764ea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "banner"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "About Us"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "<p>Company content</p>"
                                    },
                                    "banner": {
                                        "description": "Image/Video (max 10MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-03-20T09:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Page created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Page"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pages/{slug}": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Get page by slug",
                "operationId": "aa83d7826bef6cee7fa7c5fedd611247",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Page slug",
                        "required": true,
                        "example": "about-us"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Page"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/pages/{id}": {
            "put": {
                "tags": [
                    "Pages"
                ],
                "summary": "Update page",
                "operationId": "0db0c90b61af20d0c9eeb5f440b91081",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Page ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "About You"
                                    },
                                    "content": {
                                        "type": "<p>This is about you update page</p>"
                                    },
                                    "banner": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Page updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Page"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pages"
                ],
                "summary": "Delete page",
                "operationId": "1d34244ac8efd74dd341c8f746a4ce30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Page ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Page": {
                "required": [
                    "title",
                    "slug",
                    "banner_type",
                    "banner_path",
                    "content",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "banner_type": {
                        "type": "string",
                        "enum": [
                            "image",
                            "video"
                        ]
                    },
                    "banner_path": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}